+++ fast_rx_test/lib/src/rx_test.dart
@@ -2,14 +2,15 @@
 
 import 'package:fast_rx/src/rx/rx.dart';
 
-/// A container for a [transform] on [rx]
+/// A container for a [transform] on an [Rx] object
 class RxTest<T extends Rx> {
-  /// The rx value
-  final T rx;
+  /// A function to generate an [Rx] object. This is a function so that the
+  /// test can be reused.
+  final T Function() generate;
 
-  /// A transformation to run on [rx]
+  /// A transformation to run on the generated [Rx] object
   final void Function(T rx) transform;
 
   /// Create an [RxTest]
-  RxTest(this.rx, this.transform);
+  RxTest(this.generate, this.transform);
 }