+++ fast_rx_test/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0
+- `RxTest` now takes an Rx generator function so that tests can be reused
+- `expectRxNotification` improvements
+- `shouldNotify` and `shouldRegister` fields are now optional
+
 ## 0.2.0
 - Made tests much more convenient to write
 - Added expectRxNotification